[modify]　モデルを機能ごとに変更
