+++ flaskdb/model/itemModel.py
@@ -1,0 +1,11 @@
+from flaskdb import db
+
+class Item(db.Model):
+    __tablename__ = "items"
+    id = db.Column(db.Integer, primary_key=True)
+    user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
+    itemname = db.Column(db.String(128), nullable=False)
+    price = db.Column(db.Integer(), nullable=False)
+
+    def __repr__(self):
+        return "<Item %r>" % self.id
