+++ flaskdb/model/userModel.py
@@ -1,0 +1,10 @@
+from flaskdb import db
+
+class User(db.Model):
+    __tablename__ = "users"
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(64), unique=True, nullable=False)
+    password = db.Column(db.String(128), nullable=False)
+
+    def __repr__(self):
+        return "<User %r>" % self.id
